agente_backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
sqlalchemy
bcrypt
pymongo 
python-jose[cryptography]
passlib[bcrypt]
openai